code refactoring + improve pipeline for logging
