+++ fbrip/link.go
@@ -1,9 +1,6 @@
 package fbrip
 
 import (
-	"bytes"
-	"io"
-	"net/http"
 	"net/url"
 )
 
@@ -14,19 +11,3 @@ func fixUrl(Url *url.URL) *url.URL {
 	return Url
 }
 
-//It checks if `cookie` is already in `slice of cookies`
-func includesCookie(cookies []*http.Cookie, cookie *http.Cookie) bool {
-	for _, c := range cookies {
-		if c.Name == cookie.Name {
-			return true
-		}
-	}
-	return false
-}
-
-// io.Reader -> []bytes
-func bodyToBytes(body io.Reader) []byte {
-	buf := new(bytes.Buffer)
-	buf.ReadFrom(body)
-	return buf.Bytes()
-}
