+++ fbrip/user_rip_test.go
@@ -1,0 +1,30 @@
+package fbrip
+
+import (
+    "testing"
+	"github.com/stretchr/testify/assert"
+)
+
+
+func TestSenseParameters(t *testing.T) {
+	user := NewUserRip("fake_email", "fake_pass")
+	assert.Equal(t, "fake_email", user.Parameters["email"])
+	assert.Equal(t, "fake_pass", user.Parameters["pass"]) 
+	assert.Equal(t, "", user.Parameters["lsd"]) 
+	assert.Equal(t, "", user.Parameters["jazoest"]) 
+	assert.Equal(t, "", user.Parameters["m_ts"]) 
+	assert.Equal(t, "", user.Parameters["li"]) 
+	assert.Equal(t, "", user.Parameters["try_number"]) 
+	assert.Equal(t, "", user.Parameters["unrecognized_tries"]) 
+	assert.Equal(t, "", user.Parameters["login"]) 
+	user.sense()
+	assert.Equal(t, "fake_email", user.Parameters["email"])
+	assert.Equal(t, "fake_pass", user.Parameters["pass"]) 
+	assert.NotEqual(t, "", user.Parameters["lsd"]) 
+	assert.NotEqual(t, "", user.Parameters["jazoest"]) 
+	assert.NotEqual(t, "", user.Parameters["m_ts"]) 
+	assert.NotEqual(t, "", user.Parameters["li"]) 
+	assert.NotEqual(t, "", user.Parameters["try_number"]) 
+	assert.NotEqual(t, "", user.Parameters["unrecognized_tries"]) 
+	assert.NotEqual(t, "", user.Parameters["login"]) 
+}
